app/requirements.txt
# General
requests

# Django
Django
django-timezone-field 
django-cors-headers
django-debug-toolbar

# Distributed tasks
celery
redis
django-celery-beat

# Databases
psycopg2-binary==2.9.1

# File control
Pillow
django-cleanup

# Prepocessing
flashtext
cyrtranslit
django-ckeditor

# Server
gunicorn==20.1.0

django-tailwind
django-browser-reload
django-ckeditor
django-cleanup

django-admin-extra-buttons

